+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Table component: added option "highlight selected row"
 - Display a comments on the waterfall chart
 - Added visualisation "risk-review"
+- Added options: "leftTitle", "rightTitle", "leftValueColor", "rightValueColor", "primitivesLibrary" for visualisation "risk-review"
+- Added text formatting for "risk-review" visualization
 
 ### Changed
 - DatePicker intervals moves with date